requirements.txt
numpy<2.0
torch>=2.0.0
torchvision
Pillow>=9.0.0
opencv-contrib-python
scipy
ultralytics
color-matcher
spandrel
safetensors
pedalboard
transformers
wordcloud
librosa
imageio-ffmpeg